+++ ChangeLog.md
@@ -20,6 +20,8 @@ See docs/process.md for more on how version tagging works.
 
 3.1.37 (in development)
 -----------------------
+- The `EM_PYTHON_MULTIPROCESSING` environment variable no longer has any effect.
+  This was added a temporary fallback but should no longer be needed. (#19224)
 - The old reverse dependency system based on `tools/deps_info.py` has been
   removed and the existing `__deps` entries in JS library files can now be used
   to express JS-to-native dependencies.  As well being more precise, and